move button into quoteContainer and below activity form
